refactor(nav): extract WheelSwoopEngine + WasdFlight from ExperimentalControls
Move the continuous wheel accumulator + the three-phase swoop zoom + the
transient zoom-undo memory into wheelSwoopEngine.js, and the held-key set +
ramped flight velocity + the per-tick yaw-projected move into wasdFlight.js.

The orchestrator keeps the input entry points (_onWheel / _onKeyDown /
_onKeyUp / _onWindowBlur) as thin routers over the two engines: _onWheel feeds
events via wheel.accumulate(); _onTick drives wheel.drain() / wasd.drain();
the key routers own the held-key set via wasd accessors. Per-tick wheel writes
commit through the write funnel (commitMove('wheel') — dispatch only); WASD
commits as a non-wheel move (invalidate zoom-undo, then dispatch). The funnel's
clearWheelMemory callback now points at wheel.clearZoomUndo, and the plan-view /
compass / focus clear sites route to it too.

Each engine owns its own scratch. Behaviour-preserving: 302/302 characterization
tests green; the clearZoomUndo break-it re-points to the wheel engine.
