+++ src/editor/lib/nav-experimental/ExperimentalControls.js
@@ -45,12 +45,7 @@ import { isStreetLevelNav, isWasdNav } from './flag.js';
 import { ModifierState } from './modifierState.js';
 import { GestureLatch } from './gestureLatch.js';
 import { SceneBounds } from './sceneBounds.js';
-import {
-  CursorAnchor,
-  isSolidFloorHit,
-  classifyHitEntity,
-  worldHitNormal
-} from './cursorAnchor.js';
+import { CursorAnchor, classifyHitEntity } from './cursorAnchor.js';
 import { RotationIndicator } from './rotationIndicator.js';
 import { TickAnimator } from './tickAnimator.js';
 import { CollisionProbe } from './collisionProbe.js';
@@ -58,18 +53,10 @@ import { GroundedState } from './groundedState.js';
 import { SituationSensor } from './situationSensor.js';
 import { CameraWriteFunnel } from './cameraWriteFunnel.js';
 import { CommittedMotionRunner } from './committedMotionRunner.js';
+import { WheelSwoopEngine } from './wheelSwoopEngine.js';
+import { WasdFlight, MOVEMENT_KEY_CODES } from './wasdFlight.js';
 import {
-  ZOOM_PER_WHEEL_TICK,
-  FOV_PER_WHEEL_TICK,
-  WHEEL_MAX_ACCUM_TICKS,
-  WHEEL_ACCUM_EPS_TICKS,
-  WHEEL_ANCHOR_DENOM_EPS_METRES,
   ROTATION_SPEED_RAD_PER_PX,
-  WASD_SPEED_HEIGHT_FACTOR,
-  WASD_MIN_SPEED,
-  WASD_MAX_SPEED,
-  WASD_RAMP_UP_MS,
-  WASD_VERTICAL_LIFT_RATE_MPS,
   PLAN_VIEW_DURATION_MS,
   LB_PAN_MAX_STEP_METRES,
   TILT_THRESHOLD_DEFAULT_DEGREES,
@@ -77,21 +64,6 @@ import {
   MAP_PIVOT_BOUNDS_RADIUS_METRES,
   MAP_PIVOT_FAR_ACCEPT_GAIN,
   WHEEL_ZOOM_LATERAL_CAP_LOWER_BOUND_METRES,
-  WHEEL_ZOOM_LATERAL_CAP_AGL_COEFF,
-  WHEEL_GROUND_REACH_CEILING_METRES,
-  FALLBACK_FORWARD_DIST,
-  SWOOP_PHASE2_ENTRY_ELEVATION_METRES,
-  SWOOP_PHASE2_EXIT_ELEVATION_METRES,
-  SWOOP_PHASE2_MAX_TICKS_PER_FRAME,
-  SWOOP_PHASE2_FLOOR_SNAP_METRES,
-  SWOOP_PHASE3_FOV_FLOOR_DEGREES,
-  SWOOP_LANDING_FOV_DEGREES,
-  SWOOP_FOV_RAMP_EXPONENT,
-  DEFAULT_MAP_FOV_DEGREES,
-  PHASE3_FOV_WIDE_CAP_DEGREES,
-  REAIM_FADE_NEAR_METRES,
-  REAIM_FADE_FAR_METRES,
-  PHASE3_REAIM_NDC_EPS,
   DEFAULT_OVERVIEW_TILT_DEGREES,
   NORTH_AXIS,
   NORTH_BEARING_FROM_MINUS_Z,
@@ -99,7 +71,6 @@ import {
   COMPASS_NORTH_TOLERANCE_DEGREES,
   COMPASS_ROTATE_STEP_DEGREES,
   EYE_MARGIN_METRES,
-  WASD_CAMERA_RADIUS_METRES,
   FALL_DURATION_MS,
   POP_TO_ROOF_DURATION_MS,
   DOUBLECLICK_STANDOFF_PULLBACK_STEP_METRES,
@@ -114,24 +85,7 @@ import {
   decideLbMode,
   decideDragModeSwitch,
   clampOrbitRadius,
-  wheelDeltaToTicks,
-  dollyFactorForTicks,
-  fovFactorForTicks,
-  cappedDollyStep,
-  levelForwardAnchor,
-  lateralCap,
-  classifySwoopTickTarget,
-  reaimWeight,
   shiftRotateStep,
-  decideSwoopPhase,
-  phase2TargetTilt,
-  phase2AscentTilt,
-  swoopLandingFov,
-  phase2HeightFrac,
-  nextZoomUndo,
-  phase2NextElevation,
-  classifyWasdStep,
-  wasdVerticalY,
   isLegitPose,
   classifyDoubleClick,
   desiredDoubleClickPose,
@@ -142,20 +96,6 @@ import { captureNavDiscovery } from '../navAnalytics.js';
 
 const DEG2RAD = Math.PI / 180;
 const RAD2DEG = 180 / Math.PI;
-
-// The held-key movement set (WASD + arrows). Shared by the keydown
-// movement branch, the keyup release path, and the WASD ↔ rotation
-// interplay edge detection.
-const MOVEMENT_KEY_CODES = new Set([
-  'KeyW',
-  'KeyA',
-  'KeyS',
-  'KeyD',
-  'ArrowUp',
-  'ArrowDown',
-  'ArrowLeft',
-  'ArrowRight'
-]);
 
 // Normalize an angle in degrees to (-180, 180].
 function normalizeDeg(deg) {
@@ -284,8 +224,17 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       get latch() {
         return self._latch;
       },
+      get cursorAnchor() {
+        return self._cursorAnchor;
+      },
       get streetLevelEnabled() {
         return self._streetLevelEnabled;
+      },
+      get tiltThreshold() {
+        return self._tiltThreshold;
+      },
+      get wheelZoomLateralCapLowerBound() {
+        return self._wheelZoomLateralCapLowerBound;
       },
       get probe() {
         return self._probe;
@@ -301,6 +250,12 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       },
       get runner() {
         return self._runner;
+      },
+      get wheel() {
+        return self._wheel;
+      },
+      get wasd() {
+        return self._wasd;
       },
       get tick() {
         return self._tick;
@@ -318,8 +273,8 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       // camera? (Deliberately distinct from the resolveContextAction busy
       // predicate, which additionally counts recovery/inactive.)
       isCameraBusy: () =>
-        self._wasdVelocity.lengthSq() > 0 ||
-        self._wheelAccum !== 0 ||
+        self._wasd.isMoving() ||
+        self._wheel.hasAccum() ||
         self._latch.isActive() ||
         self._tick.isAnimating()
     };
@@ -334,17 +289,24 @@ export class ExperimentalControls extends THREE.EventDispatcher {
     this._sensor = new SituationSensor(this._ctx);
     // Camera-write funnel (M1): the single `change`-dispatch + wheel-memory-
     // invalidation edge every camera move passes through. `clearWheelMemory`
-    // points at the wheel's zoom-undo reset, which still lives on this class
-    // until the wheel engine extracts (a one-line re-point then).
+    // points at the wheel engine's zoom-undo reset.
     this._funnel = new CameraWriteFunnel({
       dispatch: this._ctx.dispatch,
-      clearWheelMemory: () => this._clearZoomUndo()
+      clearWheelMemory: () => this._wheel.clearZoomUndo()
     });
     // Committed-motion runner (M2): the single home for every camera-owning
     // tween (recovery ease-back, teleport, and the four preset motions) — the
     // ownership flags, the anti-stranding cancel, the per-tick write-funnel
     // commit, and the parameterized settle epilogue.
     this._runner = new CommittedMotionRunner(this._ctx);
+    // Wheel-swoop engine: the continuous wheel accumulator + the three-phase
+    // swoop zoom + the transient zoom-undo memory. Fed by the `_onWheel` router,
+    // drained each frame by `_onTick`.
+    this._wheel = new WheelSwoopEngine(this._ctx);
+    // Held-key WASD flight controller: the held-key set + ramped velocity + the
+    // per-tick yaw-projected move with the step classifier. Fed by the key
+    // routers, driven each frame by `_onTick`.
+    this._wasd = new WasdFlight(this._ctx);
 
     // TASK-010 (D2): the single tilt threshold T governing the LB
     // sub-mode, the wheel cut, the rotation regime, and the letterbox.
@@ -427,95 +389,6 @@ export class ExperimentalControls extends THREE.EventDispatcher {
     this._raycaster = new THREE.Raycaster();
     this._tmpNDC = new THREE.Vector2();
 
-    // TASK-014a (#6 Option B): continuous wheel accumulator, in signed
-    // fractional "nominal ticks" (replaces the old integer deltaY-unit
-    // _wheelBudget). _onWheel normalises each event via wheelDeltaToTicks
-    // and adds it here (clamped to ±WHEEL_MAX_ACCUM_TICKS); _drainWheel
-    // applies the high/FOV regimes as one continuous step per frame and the
-    // swoop as whole ticks under its rate-cap, carrying any sub-tick
-    // remainder frame-to-frame.
-    this._wheelAccum = 0;
-
-    // Phase 3 swoop state.
-    //
-    //   _zoomUndo — TASK-022 transient zoom-undo memory {valid, tilt, fov}.
-    //     `tilt` is the entry tilt captured at the wheel Phase-1→2 downward
-    //     crossing; `valid` gates whether the swoop-OUT reverses to it (true)
-    //     or eases to the default overview (false). Cleared by any ACTUAL
-    //     non-wheel camera move (not by a bare input event). Wheel activity
-    //     (in or out) preserves it. `fov` is reserved for TASK-014b (the
-    //     landing-FOV ascent — captured/cleared on the SAME flag; this task
-    //     writes it on capture and reads nothing from it: the C4 seam is live
-    //     but inert until 014b). Init `valid:false` so a session that opens
-    //     inside the band eases to the default overview on the first
-    //     swoop-out, not to a meaningless live tilt read as if it were a real
-    //     entry. Mutated only via the nextZoomUndo reducer (never poke fields).
-    //   TASK-027 Part A: the old latched `_phase3FovBaseline` is GONE. The
-    //     street-level wide-FOV cap is now the constant PHASE3_FOV_WIDE_CAP_
-    //     DEGREES (= min(landing, distortion)); the landing FOV is reached by a
-    //     height-driven ramp in Phase 2, not a latch on the floor crossing.
-    // See claude/specs/001-phase-3-plan.md.
-    this._zoomUndo = {
-      valid: false,
-      tilt: cameraTiltDegrees(camera),
-      fov: camera.fov
-    };
-
-    // TASK-027 Part B: cursor-lock re-aim baseline session, captured lazily on
-    // the first Phase-3 wheel tick and re-captured when the cursor pixel moves.
-    // { baselineQuat, baselineFov, ndc } | null. Cleared on leaving Phase 3 and
-    // on any non-wheel camera move (_clearZoomUndo).
-    this._phase3Reaim = null;
-
-    // TASK-027 Part C: the regime the last Phase-2 zoom-in tick resolved to
-    // ('swoop' | 'dolly' | null). A swoop↔dolly switch clears the zoom-undo
-    // memory (C2 for mixed descents). Reset to null whenever the dispatch
-    // leaves the Phase-2 band, so a stale value can't spuriously clear a fresh
-    // descent's memory.
-    this._lastSwoopRegime = null;
-
-    // TASK-027 Part C-add-2 (live-test "B"): depth (in nominal ticks, a float
-    // under TASK-014a's continuous model) of the current break-out dolly
-    // excursion — how far the camera has dollied toward a wall since the last
-    // swoop tick. Zoom-out unwinds this back to 0 (dolly back to the rail)
-    // before resuming the swoop ascent. Reset on a swoop tick / leaving the
-    // band / non-wheel move.
-    this._breakoutDollyDepth = 0;
-
-    // TASK-022 / TASK-027 Part A (L4): swoop-OUT ascent anchor, the sole stored
-    // ascent state — the three fields bundled into ONE struct `{ frac, tilt,
-    // fov } | null` so they cannot desync. null = no ascent in progress.
-    // Captured TOGETHER on the first zoom-out tick of an ascent (under the
-    // `== null` guard), nulled together on any descent tick / Phase-3 entry /
-    // ceiling hand-off / swoop↔dolly regime switch, so the next ascent
-    // re-captures from the live pose. The ascent TARGET (tilt + fov) is NOT
-    // stored — recomputed per-tick from `_zoomUndo.valid` (safe: the wheel path
-    // never flips `valid` mid-ascent).
-    this._ascentAnchor = null;
-
-    // Per-pass snapshot of the ground height directly below the camera,
-    // set at the top of each _drainWheel pass from _collisionFloorAt()
-    // (TASK-024 — the collision floor, so the swoop lands on roofs).
-    // Read (not re-probed) by _decideWheelRegime / _applyContinuousHighStep /
-    // _applyPhase2WheelTick / the phase helpers so every step in a pass —
-    // including the recursive swoop ↔ high hand-offs — sees the same ground.
-    // Distinct from
-    // _lastGroundY (the persisted miss-fallback cache). (TASK-013)
-    this._frameGroundY = 0;
-    // TASK-024a (solid-geometry guard): did this pass's ground probe HIT a
-    // real surface, or did it miss (source 'cache' = stale last-known ground,
-    // no surface below — outside a finite scene's bounds)? On a miss every
-    // swoop ground-relative clamp / phase-snap is suppressed so the wheel acts
-    // as a plain anchored dolly (free descent), letting the camera reach
-    // street level from outside the bounds.
-    this._frameGroundHit = false;
-
-    // WASD held-key set; drained per tick.
-    this._heldKeys = new Set();
-    // Current WASD velocity in world horizontal plane (m/s). Ramps up
-    // toward the target while keys are held; snaps to zero on release.
-    this._wasdVelocity = new THREE.Vector3();
-
     // ActionBar zoom-in/out hold-down intervals.
     this._zoomInInterval = null;
     this._zoomOutInterval = null;
@@ -539,15 +412,6 @@ export class ExperimentalControls extends THREE.EventDispatcher {
     // value emitted via `nav-experimental:modechange` for the LB sub-
     // mode comparator. Initialised lazily on the first gesture / tween.
     this._currentLbMode = null;
-
-    // TASK-024 recovery state.
-    //   _lastWasdBlocked — WASD block hysteresis carry (WE-3b).
-    // The camera-tween ownership flags (`_recoveryActive` / `_teleportActive`)
-    // live on the CommittedMotionRunner (this._runner); callers read them via
-    // `_tweenOwnsCamera()` / `_runner.isRecovering()`. (The legit-pose snapshot
-    // and the discoverability-cue shown state live in the SituationSensor
-    // service.)
-    this._lastWasdBlocked = false;
 
     // The per-tick context snapshot the view-button resolver reads lives in the
     // SituationSensor service (a pure read — the resolver never probes, so the
@@ -846,7 +710,7 @@ export class ExperimentalControls extends THREE.EventDispatcher {
     this._wasdEnabled = enabled;
     // Flipped off mid-flight: drop any held movement keys so the camera
     // doesn't keep flying on keys whose keyups will now be ignored.
-    if (!enabled) this._heldKeys.clear();
+    if (!enabled) this._wasd.clearHeldKeys();
   }
 
   // TASK-010 (D-LT-3 / #6): live-tunable Shift+LB rotation speed
@@ -959,14 +823,14 @@ export class ExperimentalControls extends THREE.EventDispatcher {
         // clear the zoom-undo memory. In onTick (idempotent) so it's gone the
         // instant the tween starts moving — a tween pre-empted at frame 0
         // never ticks, correctly leaving the memory intact.
-        this._clearZoomUndo();
+        this._wheel.clearZoomUndo();
         this.dispatchEvent(this._changeEvent);
       },
       onDone: () => {
         camera.position.copy(endPos);
         camera.quaternion.copy(endQuat);
         camera.updateMatrixWorld();
-        this._clearZoomUndo(); // TASK-022 (idempotent; closes the onTick window)
+        this._wheel.clearZoomUndo(); // TASK-022 (idempotent; closes the onTick window)
         this._planViewActive = false;
         this._planViewHandle = null;
         // TASK-024 (D4): reseed the legit-pose snapshot from the committed
@@ -1151,7 +1015,7 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       camera.updateMatrixWorld();
       // TASK-022: compass rotate / align / body-click top-down all route here
       // — a non-wheel camera move. Clear the zoom-undo memory.
-      this._clearZoomUndo();
+      this._wheel.clearZoomUndo();
       this.dispatchEvent(this._changeEvent);
       // this.center: the orbit pivot, or the ground point under the camera
       // for a spin/align in place. Downstream wheel-zoom references it, so
@@ -1180,7 +1044,7 @@ export class ExperimentalControls extends THREE.EventDispatcher {
         }
         camera.updateMatrixWorld();
         // TASK-022: clear the instant the tween starts moving (idempotent).
-        this._clearZoomUndo();
+        this._wheel.clearZoomUndo();
         this.dispatchEvent(this._changeEvent);
       },
       onDone: finalize
@@ -1255,7 +1119,7 @@ export class ExperimentalControls extends THREE.EventDispatcher {
         // TASK-022: focus-to-object moves the camera via the A-Frame
         // focus-animation component — a non-wheel move. Clear on its change
         // hook (fires each frame of the transition; idempotent).
-        this._clearZoomUndo();
+        this._wheel.clearZoomUndo();
         // TASK-024 (D4): reseed the legit-pose snapshot once the focus
         // (double-click teleport) animation has settled, so recovery can't
         // ease back to the pre-teleport pose. The component sets
@@ -1982,50 +1846,18 @@ export class ExperimentalControls extends THREE.EventDispatcher {
 
   _onWheel(event) {
     // TASK-012 (M-3): a camera-owning tween (recovery or teleport) owns the
-    // camera — passive wheel input is ignored, not raced (L-3).
+    // camera — passive wheel input is ignored, not raced (L-3). Input plumbing
+    // stays on the orchestrator; the accumulate + drain logic lives on the
+    // wheel-swoop engine.
     if (this._isInactive() || this._tweenOwnsCamera()) return;
     event.preventDefault();
-
     // Feature-discovery: first wheel zoom this session.
     captureNavDiscovery('zoom');
-
-    // TASK-014a (#6 Option B): accumulate only — apply no motion here (the
-    // drain owns motion + recovery suppression, exactly as before). Normalise
-    // the event to a signed fractional "nominal tick" count (deltaMode-aware,
-    // per-event clamped) and add it to the continuous accumulator.
-    const viewportH =
-      typeof window !== 'undefined' && window.innerHeight
-        ? window.innerHeight
-        : 800;
-    this._wheelAccum += wheelDeltaToTicks(
-      event.deltaY,
-      event.deltaMode,
-      viewportH
-    );
-    // A4: bound the accumulator (replaces the old WHEEL_MAX_BUDGET clamp, in
-    // the same place). High/FOV drain the whole accumulator each frame so they
-    // can't pile up, but the swoop drains only a few ticks/frame — without this
-    // a sustained fast scroll would build a runaway tail that keeps descending
-    // long after the input stops.
-    if (this._wheelAccum > WHEEL_MAX_ACCUM_TICKS) {
-      this._wheelAccum = WHEEL_MAX_ACCUM_TICKS;
-    } else if (this._wheelAccum < -WHEEL_MAX_ACCUM_TICKS) {
-      this._wheelAccum = -WHEEL_MAX_ACCUM_TICKS;
-    }
-
-    // Latest cursor position is needed at drain time; remember it.
-    this._lastWheelClientX = event.clientX;
-    this._lastWheelClientY = event.clientY;
-
-    // Ctrl+wheel = fixed-tilt zoom escape hatch (Phase 3 plan, Open
-    // Decision #2). Mac trackpad pinch arrives as Ctrl+wheel naturally,
-    // so the same code path handles pinch-to-zoom. Latch the flag at
-    // event time; drain-time consumes it per tick.
-    this._lastWheelCtrlKey = !!event.ctrlKey;
+    this._wheel.accumulate(event);
   }
 
   _onWindowBlur() {
-    this._heldKeys.clear();
+    this._wasd.clearHeldKeys();
     if (this._latch.isActive()) {
       this._latch.end();
       this._emitModeChange(null);
@@ -2102,8 +1934,8 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       // set so auto-repeat keydowns and additional movement keys do NOT
       // end a rotation started while already moving — only the WASD-mode
       // boundary does (the matching exit edge lives in `_onKeyUp`).
-      if (this._heldKeys.size === 0) this._endRotationGestureForWasd();
-      this._heldKeys.add(k);
+      if (this._wasd.noHeldKeys()) this._endRotationGestureForWasd();
+      this._wasd.addHeldKey(k);
       // Prevent the browser from scrolling the page (arrow keys) or
       // shifting focus in scrollable panels while driving the camera.
       event.preventDefault();
@@ -2549,8 +2381,8 @@ export class ExperimentalControls extends THREE.EventDispatcher {
     // TASK-010 (B6): symmetric with `_onKeyDown` — same first-line sync.
     this._syncDragModeToShift(event.shiftKey);
     const k = event.code;
-    const wasHeld = this._heldKeys.has(k);
-    if (wasHeld) this._heldKeys.delete(k);
+    const wasHeld = this._wasd.hasHeldKey(k);
+    if (wasHeld) this._wasd.deleteHeldKey(k);
     // Interplay: releasing a held movement key ends an in-progress
     // rotation gesture — functionally equivalent to Shift-up / button-up,
     // even though the user may keep dragging. `_heldKeys` only ever holds
@@ -2613,8 +2445,8 @@ export class ExperimentalControls extends THREE.EventDispatcher {
       // itself runs as a separate TickAnimator subscription via animate().)
       return;
     }
-    this._drainWheel();
-    this._drainWASD(deltaMs);
+    this._wheel.drain();
+    this._wasd.drain(deltaMs);
     // TASK-024 (3b/3e): legit-pose snapshot + discoverability cue. Runs
     // after the drains so it captures the post-move pose. Suppressed while a
     // recovery OR teleport tween owns the camera (D2 / TASK-012 M-C) — the
@@ -2630,1083 +2462,6 @@ export class ExperimentalControls extends THREE.EventDispatcher {
   // here; `_onTick` invokes it so the patch still intercepts.
   _updateLegitSnapshotAndCue() {
     this._sensor.update();
-  }
-
-  // TASK-014a (#6 Option B): continuous single-drain. ONE frame, ONE ground
-  // snapshot, ONE net-vertical bracket, ONE recovery guard — all exactly as
-  // before. The change vs the old budget drain: the high & FOV regimes apply
-  // the WHOLE pending accumulator as a single continuous step (no
-  // quantisation, no multi-frame lag), while the swoop still consumes whole
-  // ticks under its per-frame rate-cap, carrying any sub-tick remainder to a
-  // later frame. The five responsibilities the TASK-024/024a safety guard
-  // depends on (floor snapshot, recovery suppression, swoop rate-cap,
-  // net-vertical un-ground/captureH bracket, active phase-boundary hand-offs)
-  // all stay in this one method.
-  _drainWheel() {
-    // TASK-024 (D2) / TASK-012 (M-3): a recovery OR teleport tween owns the
-    // camera — drop queued wheel.
-    if (this._tweenOwnsCamera()) {
-      this._wheelAccum = 0;
-      return;
-    }
-    if (this._wheelAccum === 0) return;
-    // Snapshot the collision floor once per pass (TASK-013 → TASK-024). Every
-    // step in the loop — including the recursive swoop ↔ high hand-offs —
-    // reads this._frameGroundY so they see a single consistent ground for the
-    // frame. The swoop reads the COLLISION floor (ground OR building roof OR
-    // tiles), so a swoop over a building lands on the roof (WE-2 / C5).
-    const frameFloor = this._probe.collisionFloorAt(
-      this._camera.position.x,
-      this._camera.position.z
-    );
-    this._frameGroundY = frameFloor.y;
-    // TASK-024a (solid-geometry guard): track whether the probe hit a real
-    // surface. On a miss (outside finite bounds) the swoop phase handlers skip
-    // every ground-relative clamp so the wheel is a plain anchored dolly.
-    this._frameGroundHit = frameFloor.source !== 'cache';
-    let changed = false;
-    // TASK-024a (1.3.2 / 2.2): capture y before the drain so the net vertical
-    // move over the whole pass drives the grounded / H edges once — covering
-    // reverse-swoop, low-tilt dolly-up, Phase-2 zoom-out and Ctrl+wheel
-    // uniformly, without scattering flags across every wheel branch.
-    const wheelStartY = this._camera.position.y;
-    // The swoop rate-cap survives Option B unchanged: a whole-tick budget,
-    // latched once at the start of the frame and held for it (re-reading per
-    // iteration would unlock extra ticks at a boundary crossing — H4 of
-    // `claude/reports/007-phase-3-plan-review.md`).
-    let swoopTicksLeft = SWOOP_PHASE2_MAX_TICKS_PER_FRAME;
-    const EPS_TICK = WHEEL_ACCUM_EPS_TICKS;
-    while (Math.abs(this._wheelAccum) >= EPS_TICK) {
-      const sign = this._wheelAccum > 0 ? 1 : -1;
-      const regime = this._decideWheelRegime();
-      if (regime === 'swoop') {
-        // TASK-027 Part C: within the swoop band, a zoom-IN craning UP at a
-        // solid wall / open sky breaks out to a cursor dolly; a zoom-OUT
-        // unwinds any such break-out excursion back to the rail (Part C-add-2
-        // "B") before resuming the swoop ascent.
-        let breakout;
-        if (sign < 0) {
-          const r = this._classifyPhase2Target();
-          this._notePhase2Regime(r);
-          breakout = r === 'dolly';
-        } else {
-          breakout = this._breakoutDollyDepth > EPS_TICK;
-        }
-        if (breakout) {
-          // Continuous break-out dolly. On the way out, cap the step at the
-          // remaining excursion depth so it lands back on the rail and the
-          // remainder re-dispatches to the swoop ascent next iteration.
-          const t =
-            sign < 0
-              ? this._wheelAccum
-              : Math.min(this._wheelAccum, this._breakoutDollyDepth);
-          const tApplied = this._applyBreakoutDolly(t);
-          if (tApplied === 0) break;
-          this._wheelAccum -= tApplied;
-          if (sign < 0) this._breakoutDollyDepth += Math.abs(tApplied);
-          else {
-            this._breakoutDollyDepth = Math.max(
-              0,
-              this._breakoutDollyDepth - Math.abs(tApplied)
-            );
-          }
-          changed = true;
-        } else {
-          // A1: only fire the swoop on a WHOLE available tick AND with rate-cap
-          // headroom. A sub-tick remainder (e.g. 0.3) must NOT drive a full
-          // whole-tick descent — carry it to a later frame instead.
-          if (swoopTicksLeft < 1 || Math.abs(this._wheelAccum) < 1) break;
-          // TASK-027 Part D: restore the entry zoom-undo memory when the swoop
-          // is reached without a Phase-1 boundary capture (the no-ground
-          // free-descent bypass). Guarded by `!valid` ⇒ a no-op once captured.
-          if (sign < 0 && !this._zoomUndo.valid) {
-            this._zoomUndo = nextZoomUndo(this._zoomUndo, {
-              type: 'wheel-in-crossing',
-              tilt: cameraTiltDegrees(this._camera),
-              fov: this._camera.fov
-            });
-          }
-          this._applyPhase2WheelTick(sign); // whole-tick internals unchanged
-          this._wheelAccum -= sign; // consume one whole tick
-          swoopTicksLeft -= 1;
-          this._breakoutDollyDepth = 0; // a swoop tick commits the excursion
-          changed = true;
-        }
-      } else {
-        // high / lowtilt / fov: leaving the swoop band — reset the Part-C
-        // excursion + regime tracker so a stale value can't spuriously clear a
-        // fresh descent's memory. Apply the ENTIRE remaining accumulator as one
-        // continuous step (a zoom-in crossing phase1→phase2 stops at the
-        // boundary; the remainder re-dispatches to the swoop). Returns the
-        // ticks actually consumed.
-        this._lastSwoopRegime = null;
-        this._breakoutDollyDepth = 0;
-        const tApplied = this._applyContinuousHighStep(
-          this._wheelAccum,
-          regime
-        );
-        if (tApplied === 0) break; // safety: no progress → stop (no spin)
-        this._wheelAccum -= tApplied;
-        changed = true;
-      }
-    }
-    if (Math.abs(this._wheelAccum) < EPS_TICK) this._wheelAccum = 0;
-    if (changed) {
-      const EPS = 1e-3;
-      if (this._camera.position.y > wheelStartY + EPS) {
-        // Net-upward pass — deliberate up-move leaves the surface (1.3.2).
-        this._groundedState.checkUngroundOnRise(wheelStartY);
-      } else if (
-        this._camera.position.y < wheelStartY - EPS &&
-        !this._groundedState.grounded
-      ) {
-        // Net-downward pass while still flying (a swoop landing this pass would
-        // have grounded us, so the `!_grounded` test excludes that case) →
-        // deliberate vertical nav: lower H (D4, 2.2).
-        this._groundedState.captureH();
-      }
-      this.dispatchEvent(this._changeEvent);
-    }
-  }
-
-  // Decide which regime the wheel is in RIGHT NOW (read each loop iteration
-  // off the current, post-step camera pose). Extracted from the old
-  // `_applyWheelTick` dispatch. Elevation-first (per H1 of
-  // `claude/reports/007-phase-3-plan-review.md`): the swoop runs regardless of
-  // tilt. TASK-014d collapsed the tilt-conditional anchor split — 'high' and
-  // 'lowtilt' both dolly toward the cursor now (the lurch is bounded by the
-  // lateral cap in the dolly step, not by switching anchor source), so the
-  // two are treated identically by the drain; the labels are retained only to
-  // mark the Ctrl / no-ground / low-tilt cases.
-  //   'swoop'   — Phase 2 pedestal+tilt band (whole-tick, rate-capped)
-  //   'fov'     — Phase 3 street-level FOV-only
-  //   'lowtilt' — Ctrl+wheel, no-ground, or live tilt ≤ threshold dolly
-  //   'high'    — cursor-anchored Phase 1 dolly
-  _decideWheelRegime() {
-    const camera = this._camera;
-    // Ctrl+wheel (incl. Mac trackpad pinch) bypasses the swoop — plain
-    // camera-Z dolly at the current tilt/elevation (Open Decision #2).
-    if (this._lastWheelCtrlKey) return 'lowtilt';
-    // Street-level mode off: never dispatch to the swoop (phase 2) or the
-    // street FOV zoom (phase 3) — the wheel is a plain anchored dolly at
-    // every height, the same behaviour Ctrl+wheel gives with the mode on.
-    if (!this._streetLevelEnabled) {
-      return cameraTiltDegrees(camera) <= this._tiltThreshold
-        ? 'lowtilt'
-        : 'high';
-    }
-    // TASK-024a (solid-geometry guard): no ground below → no swoop floor to
-    // land on. Plain anchored dolly at the current tilt, never Phase 2/3.
-    if (!this._frameGroundHit) {
-      return cameraTiltDegrees(camera) <= this._tiltThreshold
-        ? 'lowtilt'
-        : 'high';
-    }
-    const yAgl = camera.position.y - this._frameGroundY;
-    let phase = decideSwoopPhase(yAgl);
-    // TASK-027 (live-test #3): float-robust / sticky street level — AGL within
-    // a sub-centimetre tolerance of the floor counts as Phase 3 ('fov'), so a
-    // rounding ulp at the just-landed boundary (`(groundY+1.5)−groundY` can
-    // round to `1.5+ulp`) can't route a street FOV-zoom-out into an immediate
-    // reverse swoop. 1 cm shift, imperceptible for swoop entry.
-    if (
-      phase === 'phase2' &&
-      yAgl <= SWOOP_PHASE2_EXIT_ELEVATION_METRES + 0.01
-    ) {
-      phase = 'phase3';
-    }
-    if (phase === 'phase2') return 'swoop';
-    if (phase === 'phase3') return 'fov';
-    // phase1: tilt-conditional split (LIVE — read instantaneous tilt).
-    return cameraTiltDegrees(camera) <= this._tiltThreshold
-      ? 'lowtilt'
-      : 'high';
-  }
-
-  // TASK-027 Part C: classify the current cursor target as a swoop landing
-  // surface or a break-out dolly target. Resolves the cursor anchor (with the
-  // additive hit normal) + camera look-direction and defers the cut to
-  // navMath.classifySwoopTickTarget.
-  _classifyPhase2Target() {
-    const hit = this._cursorAnchor.worldPointAt(
-      this._lastWheelClientX,
-      this._lastWheelClientY,
-      { maxGroundDist: WHEEL_GROUND_REACH_CEILING_METRES }
-    );
-    const isSolidFloor =
-      hit.source === 'mesh' ? isSolidFloorHit(hit.raw) : true;
-    // Break out only when craning UP at a wall/sky — looking down/level always
-    // swoops (Part C-add-1). Tilt < 0 ⇒ looking above horizontal.
-    const lookingUp = cameraTiltDegrees(this._camera) < 0;
-    return classifySwoopTickTarget({
-      source: hit.source,
-      normalY: hit.normal ? hit.normal.y : null,
-      isSolidFloor,
-      lookingUp
-    });
-  }
-
-  // TASK-027 Part C.3: a swoop↔dolly regime switch mid-descent is an intent
-  // change — invalidate the transient zoom-undo memory (C2) and drop any
-  // ascent anchor. Per-tick; no latched mode.
-  _notePhase2Regime(regime) {
-    if (this._lastSwoopRegime != null && regime !== this._lastSwoopRegime) {
-      this._clearZoomUndo();
-      this._ascentAnchor = null;
-    }
-    this._lastSwoopRegime = regime;
-  }
-
-  // TASK-027 Part C-add-2: one continuous break-out dolly step of `t` nominal
-  // ticks (the same cursor-anchored dolly as Phase 1, but WITHOUT the
-  // phase1→phase2 entry-boundary clamp — we are already inside the band). The
-  // caller tracks the excursion depth and caps the unwind. Returns the ticks
-  // applied (== t, or 0 if no cursor latch / no real anchor and vertical sky).
-  _applyBreakoutDolly(t) {
-    const camera = this._camera;
-    const x = this._lastWheelClientX;
-    const y = this._lastWheelClientY;
-    if (x == null || y == null) return t; // no cursor latch → consume, no-op
-    let hit = this._cursorAnchor.worldPointAt(x, y, {
-      maxGroundDist: WHEEL_GROUND_REACH_CEILING_METRES
-    });
-    if (hit.source !== 'mesh' && hit.source !== 'ground') {
-      hit = levelForwardAnchor(camera, FALLBACK_FORWARD_DIST);
-      if (hit == null) return t; // near-vertical at sky → consume, no move
-    }
-    this._dollyAlongRay(dollyFactorForTicks(t, ZOOM_PER_WHEEL_TICK), hit);
-    return t;
-  }
-
-  // TASK-014a (#6 Option B): apply `t` nominal ticks of high/lowtilt/FOV zoom
-  // as ONE continuous step. Returns the ticks actually consumed (== `t` for
-  // the interior case; a partial value when a boundary or clamp is hit so the
-  // loop re-dispatches the remainder). `regime` is one of 'high' | 'lowtilt'
-  // | 'fov' from `_decideWheelRegime`.
-  //
-  // Boundary handling (zoom-in crossing AGL 20 downward): clamp to the
-  // Phase-2 entry and capture TASK-022's transient zoom-undo memory (entry
-  // tilt + fov) via `nextZoomUndo`, exactly as the pre-Option-B
-  // `_applyPhase1WheelTick` did — Phase 2's descent lerp reads `_zoomUndo.tilt`.
-  _applyContinuousHighStep(t, regime) {
-    const camera = this._camera;
-    const sign = t > 0 ? 1 : -1;
-
-    if (regime === 'fov') {
-      // Phase 3 — FOV zoom at street level (TASK-027 Parts A + B; continuous
-      // per TASK-014a). Part A: the wide end is the constant
-      // PHASE3_FOV_WIDE_CAP_DEGREES (no per-entry latched baseline). Part B:
-      // the world point under the cursor is PINNED as FOV changes by re-aiming
-      // the camera (_applyPhase3Reaim). A2: the return rule is split by sign so
-      // a remainder is never left stuck in the FOV regime (which would spin).
-      const fovBefore = camera.fov; // H1: snapshot BEFORE any mutation
-      const cap = PHASE3_FOV_WIDE_CAP_DEGREES;
-      const floor = SWOOP_PHASE3_FOV_FLOOR_DEGREES;
-      // Zoom-out at/above the wide cap → ACTIVE hand-off to the swoop (consume
-      // ONE whole tick via the swoop kick-start; just returning the remainder
-      // would re-dispatch to 'fov' forever since camera.y hasn't changed). Ends
-      // the re-aim session (leaving Phase 3 upward).
-      if (sign > 0 && camera.fov >= cap - 1e-6) {
-        this._phase3Reaim = null;
-        this._applyPhase2WheelTick(sign); // whole-tick swoop kick-start
-        return sign;
-      }
-      let fov = camera.fov * fovFactorForTicks(t, FOV_PER_WHEEL_TICK);
-      if (fov < floor) fov = floor;
-      if (sign > 0 && fov > cap) fov = cap;
-      camera.fov = fov;
-      camera.updateProjectionMatrix();
-      // Part B re-aim (skip on Ctrl, or a floored/capped no-op step).
-      if (!this._lastWheelCtrlKey && fov !== fovBefore) {
-        this._applyPhase3Reaim(fovBefore);
-      }
-      // Interior FOV step, or zoom-in pinned at the 15° floor. Consume the
-      // ENTIRE remaining `t` so the loop terminates rather than spinning.
-      return t;
-    }
-
-    // Dolly. TASK-014d collapsed the tilt-conditional anchor split: cursor-
-    // anchor at EVERY tilt (the lurch is bounded by the lateral cap in
-    // `_dollyAlongRay`, not by switching anchor source). Anchor dispatch on the
-    // hit *source*: mesh/ground → a real target; fallback (open sky) → a
-    // LEVEL-forward anchor so zoom-in advances forward at constant height
-    // rather than drifting up into empty sky; near-vertical-at-sky → no move.
-    const x = this._lastWheelClientX;
-    const y = this._lastWheelClientY;
-    if (x == null || y == null) return t; // no cursor latch → consume, no-op
-    let hit = this._cursorAnchor.worldPointAt(x, y, {
-      maxGroundDist: WHEEL_GROUND_REACH_CEILING_METRES
-    });
-    if (hit.source !== 'mesh' && hit.source !== 'ground') {
-      hit = levelForwardAnchor(camera, FALLBACK_FORWARD_DIST);
-      if (hit == null) return t; // near-vertical at sky → consume, no move
-    }
-
-    const groundY = this._frameGroundY;
-    const yEntry = SWOOP_PHASE2_ENTRY_ELEVATION_METRES;
-
-    // Boundary-aware zoom-in: if the full step would drop AGL below the
-    // Phase-2 entry (and there IS a ground), stop exactly at the boundary and
-    // hand the remainder to the swoop. TASK-014d: Ctrl+wheel is the swoop
-    // BYPASS escape hatch — a plain cursor dolly at the current tilt that may
-    // descend past AGL 20 without entering the swoop, so skip the boundary when
-    // Ctrl is held. Street-level mode off: same bypass — there is no swoop to
-    // hand off to, so the dolly descends freely.
-    if (
-      sign < 0 &&
-      this._frameGroundHit &&
-      !this._lastWheelCtrlKey &&
-      this._streetLevelEnabled
-    ) {
-      const denom = camera.position.y - hit.y;
-      const targetY = groundY + yEntry;
-      // Would the full step land below the entry boundary?
-      const fullFactor = dollyFactorForTicks(t, ZOOM_PER_WHEEL_TICK);
-      const fullY = hit.y + fullFactor * denom;
-      if (fullY < targetY) {
-        // A3: degenerate denominator (near-horizontal anchor ≈ camera height)
-        // — the analytic solve divides by ~0. Fall back to the proven
-        // per-tick path: apply the full step, then post-step y-clamp exactly
-        // as the old _applyPhase1WheelTick did, and consume the whole `t`.
-        if (Math.abs(denom) <= WHEEL_ANCHOR_DENOM_EPS_METRES) {
-          this._dollyAlongRay(fullFactor, hit);
-          if (camera.position.y - groundY < yEntry) {
-            camera.position.y = targetY;
-            this._zoomUndo = nextZoomUndo(this._zoomUndo, {
-              type: 'wheel-in-crossing',
-              tilt: cameraTiltDegrees(camera),
-              fov: camera.fov
-            });
-            camera.updateMatrixWorld();
-          }
-          return t;
-        }
-        // Solve for the tick fraction t* that lands AGL exactly at the entry:
-        //   factor* = (groundY + yEntry − hit.y) / (cam.y − hit.y)
-        //   t*      = −ln(factor*) / ln(1 − α)
-        const factorStar = (targetY - hit.y) / denom;
-        // factorStar should be in (0,1) for a normal descent toward a lower
-        // anchor; guard against a non-positive (numerically degenerate) value.
-        if (factorStar > 0 && factorStar < 1) {
-          const alpha = ZOOM_PER_WHEEL_TICK;
-          const tStar = -Math.log(factorStar) / Math.log(1 - alpha);
-          this._dollyAlongRay(factorStar, hit);
-          camera.position.y = targetY; // exact y-clamp (matches old behaviour)
-          this._zoomUndo = nextZoomUndo(this._zoomUndo, {
-            type: 'wheel-in-crossing',
-            tilt: cameraTiltDegrees(camera),
-            fov: camera.fov
-          });
-          camera.updateMatrixWorld();
-          return tStar; // remainder (t − tStar) re-dispatches to the swoop
-        }
-        // Degenerate factor* — fall back to the full step + post-step clamp.
-        this._dollyAlongRay(fullFactor, hit);
-        if (camera.position.y - groundY < yEntry) {
-          camera.position.y = targetY;
-          this._zoomUndo = nextZoomUndo(this._zoomUndo, {
-            type: 'wheel-in-crossing',
-            tilt: cameraTiltDegrees(camera),
-            fov: camera.fov
-          });
-          camera.updateMatrixWorld();
-        }
-        return t;
-      }
-    }
-
-    // Interior step (no boundary crossing, or free descent with no ground):
-    // apply the full continuous dolly and consume the whole `t`.
-    this._dollyAlongRay(dollyFactorForTicks(t, ZOOM_PER_WHEEL_TICK), hit);
-    return t;
-  }
-
-  // Translate the camera along the camera→hit ray by the continuous `factor`
-  // (factor < 1 = closer; > 1 = farther), with the HORIZONTAL component of the
-  // translation capped (TASK-014d, via cappedDollyStep). The cap scales the
-  // whole step vector uniformly, so the move stays on the camera→hit ray
-  // (target stays under the cursor) and reversibility about a fixed target is
-  // exact. `factor` is the continuous generalisation of the old per-tick step
-  // (dollyFactorForTicks(t)·dollyFactorForTicks(−t) === 1). A non-finite step
-  // (degenerate grazing ray) is dropped — a no-op rather than NaN-ing the
-  // camera. TASK-027 Part F: the cap scales with height — max(lowerBound,
-  // 0.1×AGL) — bounding the lurch proportionally; falls to the lower bound on
-  // the no-AGL path (Ctrl+wheel / out of bounds, where AGL is non-finite).
-  _dollyAlongRay(factor, hit) {
-    const camera = this._camera;
-    const yAgl = this._frameGroundHit
-      ? camera.position.y - this._frameGroundY
-      : NaN;
-    const cap = lateralCap(
-      yAgl,
-      this._wheelZoomLateralCapLowerBound,
-      WHEEL_ZOOM_LATERAL_CAP_AGL_COEFF
-    );
-    const newPos = cappedDollyStep({
-      camPos: camera.position,
-      hit,
-      factor,
-      lateralCapMetres: cap
-    });
-    if (newPos == null) return; // non-finite step: skip this tick
-    camera.position.copy(newPos);
-
-    // Track far plane based on distance.
-    const distance = camera.position.distanceTo(this.center);
-    camera.far = Math.min(100000000, Math.max(20000, distance * 10));
-    camera.updateProjectionMatrix();
-    camera.updateMatrixWorld();
-  }
-
-  // Whole-tick Phase 1 / low-tilt dolly used by the swoop's active hand-offs
-  // (Phase 2 → Phase 1 zoom-out). These only ever hand off one whole tick's
-  // worth, in the same drain pass that latched the cursor — so they keep the
-  // pre-Option-B whole-tick form (sidesteps M7: a fractional swoop-exit anchor
-  // never arises). Routes to the same `_dollyAlongRay` math the continuous
-  // step uses, at the per-whole-tick factor.
-  _applyPhase1WheelTick(sign) {
-    const camera = this._camera;
-    const x = this._lastWheelClientX;
-    const y = this._lastWheelClientY;
-    if (x == null || y == null) return;
-    // TASK-014d collapsed anchor dispatch (same as the continuous step):
-    // cursor at every tilt, level-forward on a no-real-hit (sky), no move when
-    // the heading is vertical-undefined.
-    let hit = this._cursorAnchor.worldPointAt(x, y, {
-      maxGroundDist: WHEEL_GROUND_REACH_CEILING_METRES
-    });
-    if (hit.source !== 'mesh' && hit.source !== 'ground') {
-      hit = levelForwardAnchor(camera, FALLBACK_FORWARD_DIST);
-      if (hit == null) return; // near-vertical at sky → no move
-    }
-    this._dollyAlongRay(dollyFactorForTicks(sign, ZOOM_PER_WHEEL_TICK), hit);
-
-    // TASK-024a (solid-geometry guard): no ground below → no Phase-2 boundary.
-    if (!this._frameGroundHit) return;
-
-    // Boundary: Phase 1 → Phase 2 on zoom-in (post-step y-clamp). TASK-022:
-    // capture the zoom-undo memory at the crossing, same as the continuous step.
-    const groundY = this._frameGroundY;
-    if (
-      sign < 0 &&
-      this._streetLevelEnabled && // mode off: no Phase-2 boundary to clamp at
-      camera.position.y - groundY < SWOOP_PHASE2_ENTRY_ELEVATION_METRES
-    ) {
-      camera.position.y = groundY + SWOOP_PHASE2_ENTRY_ELEVATION_METRES;
-      this._zoomUndo = nextZoomUndo(this._zoomUndo, {
-        type: 'wheel-in-crossing',
-        tilt: cameraTiltDegrees(camera),
-        fov: camera.fov
-      });
-      camera.updateMatrixWorld();
-    }
-  }
-
-  // Phase 2 — pedestal + tilt-toward-horizontal. No cursor anchoring
-  // (see plan §"Design history note"). Yaw and (x,z) preserved across
-  // the tick; only y and tilt change.
-  //
-  // Boundary handling at zoom-out is *active*, not deferred to next
-  // tick. The naive "clamp at boundary, let next tick re-dispatch"
-  // model deadlocks because `decideSwoopPhase(yCeil)` returns 'phase2'
-  // (the table is inclusive on the Phase 2 side at y = yCeil) and the
-  // next tick fires the boundary again. The active hand-off applies
-  // this tick's energy in the destination phase. Found at feel-test
-  // 2026-05-11.
-  //
-  // Runs entirely in AGL space (yAgl = camera.y − groundY), reading the
-  // per-pass ground snapshot `this._frameGroundY`, and writes the result
-  // back as absolute camera.y = groundY + yAglNext (TASK-013). On a flat
-  // scene at y=0 this is behaviour-identical to the old absolute math
-  // (modulo one cheap extra probe per pass).
-  //
-  // Boundary handling (all in AGL):
-  //   zoom-in: yAglNext ≤ yFloor → snap to floor, tilt to 0°, FOV eased to the
-  //     landing FOV (Part A). Next tick dispatches naturally to Phase 3.
-  //   zoom-in: yAglNext within SWOOP_PHASE2_FLOOR_SNAP_METRES of yFloor →
-  //     snap (H6).
-  //   zoom-out: yAgl in [0, yFloor + snap] → kick-start to yFloor + snap.
-  //     The multiplicative reciprocal `1.5 + (yAgl-1.5)/(1-α)` is zero at
-  //     yAgl=yFloor exactly, so without the kick-start zoom-out from
-  //     street level produces no motion. The `yAgl >= 0` lower bound
-  //     suppresses a stale-cache teleport (negative AGL — see below)
-  //     while preserving every legitimate fresh-probe kick-start.
-  //   zoom-out: yAglNext ≥ yCeil → clamp to groundY + yCeil, set tilt to
-  //     the recomputed ascentTarget, hand the tick's energy to Phase 1.
-  //
-  // TASK-022 (model 1D): the DESCENT (zoom-in) tilt is unchanged — it lerps
-  // the captured entry tilt (`_zoomUndo.tilt`) toward 0° via phase2TargetTilt.
-  // The ASCENT (zoom-out) interpolates from the camera's LIVE current tilt
-  // toward a target — the captured entry tilt if `_zoomUndo.valid`, else the
-  // DEFAULT_OVERVIEW_TILT_DEGREES default — anchored at (startFrac, startTilt)
-  // captured on the ascent's first tick (no jump, WE-5; exact reverse for an
-  // immediate undo, C1). Both legs go through the roll-safe re-tilt below.
-  _applyPhase2WheelTick(sign) {
-    const camera = this._camera;
-    const groundY = this._frameGroundY; // pass snapshot
-    const yFloor = SWOOP_PHASE2_EXIT_ELEVATION_METRES; // AGL floor 1.5
-    const yCeil = SWOOP_PHASE2_ENTRY_ELEVATION_METRES; // AGL ceil 20
-    const snap = SWOOP_PHASE2_FLOOR_SNAP_METRES; // 1.0
-
-    let yAgl = camera.position.y - groundY; // ← convert in
-
-    // TASK-022: the swoop-OUT target, recomputed per-tick (NOT stored — the
-    // wheel path never flips `_zoomUndo.valid` mid-ascent, so it is constant
-    // across an ascent whether read once or per-tick). Memory valid → reverse
-    // to the captured entry tilt; else ease to the default overview.
-    const ascentTarget = this._zoomUndo.valid
-      ? this._zoomUndo.tilt
-      : DEFAULT_OVERVIEW_TILT_DEGREES;
-    // TASK-027 Part A: the swoop-OUT FOV target mirrors the tilt — memory valid
-    // → exact FOV undo (the captured entry FOV); else the default map FOV.
-    const ascentTargetFov = this._zoomUndo.valid
-      ? this._zoomUndo.fov
-      : DEFAULT_MAP_FOV_DEGREES;
-
-    if (sign > 0) {
-      // ASCENT (zoom-out): atomically capture the ascent anchor on the FIRST
-      // out-tick of this ascent (the sole writer, under the == null guard).
-      // Only the TILT needs an anchor (the user can crane the camera mid-swoop,
-      // so the ascent tilt must start from the live pose without a jump). FOV
-      // can't be perturbed mid-band (only the wheel changes it), so the ascent
-      // FOV is a pure function of height (swoopLandingFov) — no anchor needed.
-      // `yAgl` here is the PRE-step height (matches the camera's live tilt read
-      // below).
-      if (this._ascentAnchor == null) {
-        this._ascentAnchor = {
-          frac: phase2HeightFrac(yAgl),
-          tilt: cameraTiltDegrees(camera)
-        };
-      }
-    } else {
-      // DESCENT (zoom-in): reset the ascent anchor so the next ascent
-      // re-captures from the live pose. Descent tilt semantics are unchanged.
-      this._ascentAnchor = null;
-    }
-
-    // Zoom-out kick-start (AGL-relative — WE-7). A FRESH downward probe
-    // always yields yAgl >= 0 (the hit ground is below the camera), so
-    // the `yAgl >= 0` lower bound preserves EVERY legitimate case —
-    // including the saved-scene-below-floor kick-start (a camera at
-    // AGL 0.5 on real ground must still kick-start). A NEGATIVE yAgl can
-    // only arise from the D2 stale cache (cached ground ABOVE the camera —
-    // camera over a gap, reachable via WASD-during-Phase-3 then
-    // zoom-out); that is exactly the teleport case to suppress.
-    if (sign > 0 && yAgl >= 0 && yAgl <= yFloor + snap) {
-      yAgl = yFloor + snap;
-    }
-
-    let yAglNext = phase2NextElevation(yAgl, sign);
-
-    // Floor snap on zoom-in (H6) — AGL-relative.
-    if (sign < 0 && yAglNext - yFloor < snap) {
-      yAglNext = yFloor;
-    }
-
-    // Boundary: Phase 2 → Phase 3 on zoom-in.
-    if (sign < 0 && yAglNext <= yFloor) {
-      camera.position.y = groundY + yFloor; // ← write back
-      this._setCameraTiltPreservingYaw(0);
-      // TASK-027 Part A: reach the landing FOV exactly at the floor (no latched
-      // baseline; the wide cap is the PHASE3_FOV_WIDE_CAP_DEGREES constant).
-      camera.fov = swoopLandingFov(
-        yFloor,
-        this._zoomUndo.fov,
-        SWOOP_LANDING_FOV_DEGREES,
-        SWOOP_FOV_RAMP_EXPONENT
-      );
-      camera.updateProjectionMatrix();
-      // TASK-022: Phase-3 entry ends any ascent geometry; reset the anchor so
-      // a fresh ascent re-captures from the live pose (already null on a
-      // descent run, but explicit at the band exit).
-      this._ascentAnchor = null;
-      camera.updateMatrixWorld();
-      // TASK-024a (D1, 1.2.2 / PA-3): the wheel swoop has no onDone — landing
-      // IS this Phase-2→3 boundary crossing. DERIVE (don't force-true): the
-      // landing height is groundY + SWOOP_PHASE2_EXIT_ELEVATION, a constant
-      // INDEPENDENT of EYE_MARGIN, so deriving runs the real ≤ eye-margin test
-      // against a fresh collision-floor probe and survives either constant
-      // being retuned. `groundY` here is the collision floor under the camera,
-      // so a swoop onto a roof grounds to the roof (D5/WE-7).
-      this._groundedState.deriveFromPose();
-      this._maybeEmitLbModeChange();
-      return;
-    }
-
-    // Boundary: Phase 2 → Phase 1 on zoom-out. Hand the tick off
-    // actively so the wheel click visibly continues past AGL=yCeil
-    // rather than deadlocking at the boundary.
-    if (sign > 0 && yAglNext >= yCeil) {
-      camera.position.y = groundY + yCeil; // ← write back
-      // TASK-022: the ceiling tilt is the recomputed ascentTarget (the
-      // captured entry tilt if memory valid, else the default overview) — the
-      // same target the per-tick ascent ramps toward. A ≤90° single-step arc,
-      // applied via the roll-safe re-tilt (antiparallel guard not triggered).
-      this._setCameraTiltPreservingYaw(ascentTarget);
-      // TASK-027 Part A: set FOV to the ascent target in one step (mirrors the
-      // tilt), so leaving the band upward always restores a sane FOV (entry FOV
-      // if memory valid, else the 60° map default) — closing the stale-FOV
-      // window where a re-descent would read a leftover-narrow `_zoomUndo.fov`.
-      camera.fov = ascentTargetFov;
-      camera.updateProjectionMatrix();
-      // TASK-022: ascent complete at the ceiling: reset the anchor. Above the
-      // ceiling the tilt is the user's to set freely (Phase 1 tilt-preserving).
-      this._ascentAnchor = null;
-      camera.updateMatrixWorld();
-      this._maybeEmitLbModeChange();
-      // Now dispatch a Phase 1 tick. TASK-014d collapsed the tilt split, so
-      // this is always the cursor-anchored Phase-1 tick (reads the same
-      // `this._frameGroundY` snapshot). This is a sign > 0 (zoom-out) tick
-      // and the Phase-1 boundary clamp body is sign < 0-gated, so routing
-      // through the full Phase-1 tick does NOT re-fire the clamp or re-latch
-      // the ascent/undo tilt state (TASK-022) — exactly one anchored dolly
-      // step happens here.
-      return this._applyPhase1WheelTick(sign);
-    }
-
-    camera.position.y = groundY + yAglNext; // ← write back
-    // TASK-022: per-tick re-tilt, branched on direction.
-    //   sign < 0 (descent): unchanged — lerp the captured entry tilt → 0°.
-    //   sign > 0 (ascent):  interpolate the ascent anchor (startFrac,
-    //     startTilt) → ascentTarget, anchored so there is no jump (WE-5) and
-    //     an immediate undo retraces the descent exactly (C1).
-    if (sign < 0) {
-      this._setCameraTiltPreservingYaw(
-        phase2TargetTilt(yAglNext, this._zoomUndo.tilt)
-      );
-      // TASK-027 Part A: descent FOV ramp — ease the entry FOV open toward the
-      // landing FOV as AGL falls, back-loaded into the final stretch.
-      camera.fov = swoopLandingFov(
-        yAglNext,
-        this._zoomUndo.fov,
-        SWOOP_LANDING_FOV_DEGREES,
-        SWOOP_FOV_RAMP_EXPONENT
-      );
-    } else {
-      this._setCameraTiltPreservingYaw(
-        phase2AscentTilt(
-          yAglNext,
-          this._ascentAnchor.frac,
-          this._ascentAnchor.tilt,
-          ascentTarget
-        )
-      );
-      // TASK-027 Part A: ascent FOV — a pure function of height (narrow =
-      // ascent target), the SAME curve the descent drew, so an immediate undo
-      // retraces it exactly; no anchor needed (FOV can't be perturbed mid-band).
-      camera.fov = swoopLandingFov(
-        yAglNext,
-        ascentTargetFov,
-        SWOOP_LANDING_FOV_DEGREES,
-        SWOOP_FOV_RAMP_EXPONENT
-      );
-    }
-    camera.updateProjectionMatrix();
-    camera.updateMatrixWorld();
-    // Toolbar visual indicator: Phase 2's tilt lerp crosses the 30°
-    // boundary silently from the LB-mode comparator's perspective. Emit
-    // here so the toolbar restyles in lock-step with the swoop. (Phase
-    // 1 and Phase 3 are tilt-preserving, so no equivalent calls needed
-    // there.)
-    this._maybeEmitLbModeChange();
-  }
-
-  // (Phase 3 FOV-only zoom is now handled continuously inside
-  // `_applyContinuousHighStep` (regime 'fov') — the old per-tick
-  // `_applyPhase3WheelTick` was folded in for TASK-014a Option B. The
-  // zoom-out→Phase-2 baseline hand-off and the lazy baseline latch live
-  // there now.)
-
-  // TASK-027 Part B — cursor-lock re-aim. Re-aims the camera so the world point
-  // under the cursor stays pinned to the same screen pixel as FOV changes. The
-  // orientation is rebuilt ABSOLUTELY from a captured baseline pose every tick
-  // (not composed incrementally), so it is a pure function of FOV → exactly
-  // reversible (WE-B2) and unwinds to the entry pose at baseline FOV (B.3). The
-  // exact ordering is load-bearing (H1/H2/H3) — do not reorder:
-  //   FOV already applied + updateProjectionMatrix (caller) →
-  //   copy baselineQuat + updateMatrixWorld → raycast cursor pixel →
-  //   slerp the minimal-arc QUATERNION by the continuity weight → premultiply.
-  _applyPhase3Reaim(fovBefore) {
-    const camera = this._camera;
-    const ndc = this._cursorAnchor.ndcFor(
-      this._lastWheelClientX,
-      this._lastWheelClientY
-    );
-
-    // Baseline session: capture on the first Phase-3 tick; re-capture at the
-    // current pose when the cursor PIXEL moves (a new aim — Δ starts at 0, no
-    // jump). The target world point `P` is resolved ONCE at capture and held
-    // for the whole session (live-test #3 fix): re-resolving the cursor target
-    // every tick breaks the unwind — once the re-aim cranes the camera, the
-    // cursor pixel points somewhere else (e.g. sky above the building), so on
-    // zoom-out it can't find the original target to un-crane back to. A stable
-    // P makes the re-aim a pure function of fov for the session → it unwinds
-    // exactly back to the baseline (street) pose as the FOV widens. (This is
-    // the notes' "first-frame-per-gesture capture"; the spec's WE-B caveat —
-    // tile streaming voids retrace — is the accepted cost.) baselineFov is the
-    // PRE-step FOV (H1).
-    if (
-      !this._phase3Reaim ||
-      ndc.distanceTo(this._phase3Reaim.ndc) > PHASE3_REAIM_NDC_EPS
-    ) {
-      const hit = this._cursorAnchor.worldPointAt(
-        this._lastWheelClientX,
-        this._lastWheelClientY,
-        { maxGroundDist: WHEEL_GROUND_REACH_CEILING_METRES }
-      );
-      this._phase3Reaim = {
-        baselineQuat: camera.quaternion.clone(),
-        baselineFov: fovBefore,
-        ndc: ndc.clone(),
-        // No real hit (open sky) at capture → no target to pin this session.
-        targetP:
-          hit.source === 'fallback'
-            ? null
-            : new THREE.Vector3(hit.x, hit.y, hit.z),
-        targetDist: hit.distance
-      };
-    }
-    const reaim = this._phase3Reaim;
-    if (!reaim.targetP) return; // sky aim → centre-anchored FOV change (B.4)
-
-    // camPos is fixed through Phase 3, so the captured P gives a stable aim.
-    const toP = this._tmpV3f
-      .subVectors(reaim.targetP, camera.position)
-      .normalize();
-    if (toP.lengthSq() < 1e-12) return; // P ≈ camera position: nothing to aim at
-
-    // (H2) Re-orient to the BASELINE quat and refresh matrixWorld BEFORE the
-    // raycast, so the cursor-pixel world ray is sampled under (baselineQuat,
-    // NEW fov) — not the previous tick's premultiplied orientation.
-    camera.quaternion.copy(reaim.baselineQuat);
-    camera.updateMatrixWorld();
-    this._reaimRaycaster.setFromCamera(reaim.ndc, camera);
-    const rayDir = this._reaimRaycaster.ray.direction; // unit, world space
-
-    // (M4) Continuity weight: fade re-aim to 0 as the target recedes toward the
-    // horizon, so the façade → sky crossing is continuous. (H3) Scale the
-    // QUATERNION via slerp from identity — never lerp the directions (that would
-    // change the axis with the weight and break reversibility).
-    const w = reaimWeight(
-      reaim.targetDist,
-      REAIM_FADE_NEAR_METRES,
-      REAIM_FADE_FAR_METRES
-    );
-    const fullArc = this._tmpQuatB.setFromUnitVectors(rayDir, toP);
-    const delta = this._tmpQuatC.identity().slerp(fullArc, w);
-    camera.quaternion.premultiply(delta);
-    camera.quaternion.normalize();
-    camera.updateMatrixWorld();
-  }
-
-  // TASK-022: invalidate the transient zoom-undo memory. Call from a site that
-  // has just committed an actual non-wheel camera move (past its own no-op
-  // early-returns AND any zero-delta gate). Idempotent (reducer returns
-  // valid:false again). TASK-027 Part C.3 adds ONE wheel-path caller: a
-  // swoop↔dolly regime switch mid-descent (`_notePhase2Regime`) is a
-  // deliberate intent change that clears the memory — the only sanctioned
-  // wheel-path call.
-  _clearZoomUndo() {
-    this._zoomUndo = nextZoomUndo(this._zoomUndo, { type: 'non-wheel-move' });
-    // TASK-027: a real non-wheel move also ends any in-flight cursor-lock
-    // re-aim session (Part B) and any Phase-2 descent regime run (Part C) —
-    // the camera is no longer where the captured baseline/regime assumed.
-    this._phase3Reaim = null;
-    this._lastSwoopRegime = null;
-    this._breakoutDollyDepth = 0;
-  }
-
-  // Apply a tilt (in degrees from horizontal, positive = looking down)
-  // while preserving the camera's current yaw. Used by Phase 2 (both swoop
-  // legs).
-  //
-  // TASK-022: re-tilt is now ROLL-SAFE and NADIR-CONTINUOUS. We build the
-  // absolute target forward from the live yaw + commanded tiltDeg (the same
-  // direction the old lookAt aimed at — so descent and ascent passing through
-  // the same height command identically-pointed forwards, C1), then rotate
-  // the camera's TRUE current forward onto it with the minimal-arc rotation
-  // and apply it via premultiply (modelled on TASK-023's _shiftRotate). The
-  // shortest-arc axis `curFwd × newFwd` lies in the yaw-tilt plane (≈ the
-  // camera's right axis), never the forward axis, so it adds NO roll — any
-  // roll the camera carries in is preserved exactly, and there is no world-up
-  // lookAt singularity at nadir (the old path's catastrophe).
-  _setCameraTiltPreservingYaw(tiltDeg) {
-    const camera = this._camera;
-    // (1) Capture the TRUE current forward BEFORE any yaw-flattening — keep it
-    //     un-flattened through the rotation build. Flattening it first would
-    //     make the "current" forward read as tilt=0 every tick, so the arc to
-    //     newFwd would pitch by the FULL tiltDeg every tick (a runaway).
-    const curFwd = this._tmpV3d;
-    camera.getWorldDirection(curFwd); // TRUE current forward — keep
-    // Current yaw from a FLATTENED COPY of the forward.
-    const fwd = this._tmpV3a;
-    fwd.copy(curFwd);
-    fwd.y = 0;
-    if (fwd.lengthSq() < 1e-6) {
-      // Camera looking straight up/down — yaw is undefined. Use camera's
-      // local +Y projected to horizontal as a stand-in (matches the
-      // WASD-degenerate convention).
-      this._tmpV3b.set(0, 1, 0).applyQuaternion(camera.quaternion);
-      this._tmpV3b.y = 0;
-      if (this._tmpV3b.lengthSq() > 1e-6) {
-        fwd.copy(this._tmpV3b).normalize();
-      } else {
-        fwd.set(0, 0, -1);
-      }
-    } else {
-      fwd.normalize();
-    }
-    // (2) Build the absolute target forward newFwd from yaw + tiltDeg.
-    const tiltRad = tiltDeg * DEG2RAD;
-    const cos = Math.cos(tiltRad);
-    const sin = Math.sin(tiltRad);
-    const newFwd = this._tmpV3c.set(fwd.x * cos, -sin, fwd.z * cos);
-    // (3) Minimal-arc rotation from the TRUE current forward onto newFwd,
-    //     applied via premultiply (drops lookAt).
-    const R = this._tmpQuat;
-    if (curFwd.dot(newFwd) < -0.9999) {
-      // Antiparallel (180° flip): setFromUnitVectors' cross product underflows
-      // and picks an arbitrary axis → unpredictable roll/flip. Choose a fixed,
-      // roll-free axis — the camera's right — explicitly. (No Phase-2 path
-      // reaches this: the per-tick step is ≤ a few °, and the largest
-      // single-step hand-off is a ≤90° arc. Guard is mandatory per the
-      // spec-review-checklist degenerate-axis requirement.)
-      const axis = this._tmpV3b.set(1, 0, 0).applyQuaternion(camera.quaternion); // camera-right in world
-      R.setFromAxisAngle(axis, Math.PI);
-    } else {
-      R.setFromUnitVectors(curFwd, newFwd);
-    }
-    camera.quaternion.premultiply(R);
-    camera.quaternion.normalize(); // drift guard, mirrors _shiftRotate (A1)
-  }
-
-  _drainWASD(deltaMs) {
-    const camera = this._camera;
-    // TASK-024 (D2) / TASK-012 (M-3): a recovery OR teleport tween owns the
-    // camera — held keys must not fight it. Snap velocity to zero and suspend.
-    if (this._tweenOwnsCamera()) {
-      this._wasdVelocity.set(0, 0, 0);
-      return;
-    }
-    const dirX =
-      this._heldKeys.has('KeyD') || this._heldKeys.has('ArrowRight') ? 1 : 0;
-    const dirXNeg =
-      this._heldKeys.has('KeyA') || this._heldKeys.has('ArrowLeft') ? 1 : 0;
-    const dirZ =
-      this._heldKeys.has('KeyW') || this._heldKeys.has('ArrowUp') ? 1 : 0;
-    const dirZNeg =
-      this._heldKeys.has('KeyS') || this._heldKeys.has('ArrowDown') ? 1 : 0;
-    const strafe = dirX - dirXNeg;
-    const fwd = dirZ - dirZNeg;
-    const hasInput = strafe !== 0 || fwd !== 0;
-
-    // Release semantics: any frame with no held movement keys snaps the
-    // velocity to zero immediately. No deceleration ramp.
-    if (!hasInput) {
-      if (this._wasdVelocity.lengthSq() === 0) return;
-      this._wasdVelocity.set(0, 0, 0);
-      return;
-    }
-
-    // Forward = horizontal projection of camera -Z, normalized; if degenerate
-    // (camera looking straight down), fall back to camera +Y horizontal projection.
-    const forward = this._tmpV3a;
-    camera.getWorldDirection(forward); // -Z direction, normalized
-    forward.y = 0;
-    if (forward.lengthSq() > 0.0001) {
-      forward.normalize();
-    } else {
-      const up = this._tmpV3c.set(0, 1, 0).applyQuaternion(camera.quaternion);
-      up.y = 0;
-      if (up.lengthSq() > 0.0001) forward.copy(up).normalize();
-      else forward.set(0, 0, -1);
-    }
-    // Right = forward × worldUp. For forward=(0,0,-1), worldUp=(0,1,0),
-    // this yields (1,0,0), which is screen-right for an upright camera.
-    const right = this._tmpV3b.copy(forward).cross(this._tmpV3c.set(0, 1, 0));
-    right.y = 0;
-    right.normalize();
-
-    // Target velocity: unit direction × height-scaled speed. Copy out of
-    // the scratch vector — the classifier raycasts below reuse _tmpV3a/b/c.
-    this._tmpV3c.set(0, 0, 0);
-    this._tmpV3c.addScaledVector(forward, fwd);
-    this._tmpV3c.addScaledVector(right, strafe);
-    this._tmpV3c.normalize();
-    const targetDirX = this._tmpV3c.x;
-    const targetDirZ = this._tmpV3c.z;
-
-    // TASK-024: WASD fly-speed scales by TRAVEL HEIGHT (height above the
-    // land/ground beneath buildings), NOT the collision floor — so speed
-    // doesn't crawl over a building roof (B4 speed rationale; TASK-013 WE-4).
-    const groundY = this._probe.travelHeightFloorYBelow();
-    const aglRaw = camera.position.y - groundY;
-    const height = Math.max(0.1, aglRaw);
-    const targetSpeed = THREE.MathUtils.clamp(
-      height * WASD_SPEED_HEIGHT_FACTOR,
-      WASD_MIN_SPEED,
-      WASD_MAX_SPEED
-    );
-    const targetVel = new THREE.Vector3(
-      targetDirX,
-      0,
-      targetDirZ
-    ).multiplyScalar(targetSpeed);
-
-    // Acceleration ramp toward target. accel = max-speed / ramp-time so a
-    // standing-start key-press reaches WASD_MAX_SPEED in WASD_RAMP_UP_MS;
-    // for lower target speeds the ramp completes proportionally faster.
-    const accel = WASD_MAX_SPEED / (WASD_RAMP_UP_MS / 1000);
-    const maxStep = accel * (deltaMs / 1000);
-    const dv = new THREE.Vector3().subVectors(targetVel, this._wasdVelocity);
-    const dvMag = dv.length();
-    if (dvMag <= maxStep) {
-      this._wasdVelocity.copy(targetVel);
-    } else {
-      this._wasdVelocity.add(dv.multiplyScalar(maxStep / dvMag));
-    }
-
-    if (this._wasdVelocity.lengthSq() === 0) return;
-    const distMetres = deltaMs / 1000;
-    const move = new THREE.Vector3()
-      .copy(this._wasdVelocity)
-      .multiplyScalar(distMetres);
-
-    // TASK-024 (2b): forward-ray step classifier. Decide block / step-up /
-    // follow / hover from the surface geometry ahead before committing the
-    // horizontal move + any y change.
-    const stepThisFrame = Math.hypot(move.x, move.z);
-    const outcome = this._classifyWasdMove(
-      targetDirX,
-      targetDirZ,
-      stepThisFrame
-    );
-
-    if (outcome.kind === 'block') {
-      // Stop at the wall: cancel the horizontal step, snap velocity to 0,
-      // don't advance the centre.
-      this._wasdVelocity.set(0, 0, 0);
-      this._lastWasdBlocked = true;
-      return;
-    }
-    this._lastWasdBlocked = false;
-
-    // Apply the horizontal move.
-    camera.position.x += move.x;
-    camera.position.z += move.z;
-    this.center.x += move.x;
-    this.center.z += move.z;
-    // TASK-022: an advancing WASD step is a non-wheel camera move → clear the
-    // zoom-undo memory. The `block` outcome early-returns above (zero-delta —
-    // no clear); a hover/step-up that changes only y still moved → clear.
-    this._clearZoomUndo();
-
-    if (outcome.kind === 'step-up' || outcome.kind === 'follow') {
-      // TASK-024a (DEC-A/DEC-B): grounded collision-follow (preserve AGL,
-      // push-up clamp — walking hugs the surface); not-grounded eases toward
-      // the option-3 absolute target `max(H, collisionFloorDest + eye)`,
-      // rate-limited per tick so the lift/settle composes with continuous WASD.
-      // Lazily seed H if we are flying but never captured it (scene loaded
-      // not-grounded). The helper only READS H. `distMetres` (== deltaMs/1000)
-      // is the per-tick dt in seconds; reuse it for the rate limit.
-      if (!this._groundedState.grounded && this._groundedState.H == null) {
-        this._groundedState.captureH();
-      }
-      const newY = wasdVerticalY({
-        grounded: this._groundedState.grounded,
-        camY: camera.position.y,
-        floorNowY: outcome.floorNowY,
-        collisionFloorDestY: outcome.floorDestY,
-        destFloorHit: outcome.destFloorHit,
-        H: this._groundedState.H,
-        eyeMargin: EYE_MARGIN_METRES,
-        dtSeconds: distMetres,
-        rateMps: WASD_VERTICAL_LIFT_RATE_MPS
-      });
-      const dy = newY - camera.position.y;
-      camera.position.y = newY;
-      this.center.y += dy;
-    } else if (outcome.kind === 'hover') {
-      // TASK-024a (1.3.3 / WE-4): walking off a sharp drop floats the camera
-      // off the surface — the ground is now far below; un-ground and capture H
-      // at the roof height so the next W holds altitude over the street rather
-      // than terrain-following down. Only the grounded→not-grounded transition
-      // matters. y itself is held (no plunge — WE-5); centre y unchanged.
-      if (this._groundedState.grounded) {
-        this._groundedState.grounded = false;
-        this._groundedState.captureH();
-      }
-    }
-
-    camera.updateMatrixWorld();
-    this.dispatchEvent(this._changeEvent);
-  }
-
-  // TASK-024 (2b): run the WASD forward-ray + destination-floor probes and
-  // classify the step. Returns { kind, floorDestY }. `kind` is
-  // 'block' | 'step-up' | 'follow' | 'hover'. Gated to held-WASD-with-input
-  // (the caller only invokes it inside the hasInput branch), so idle frames
-  // cost nothing.
-  _classifyWasdMove(dirX, dirZ, stepThisFrame) {
-    const camera = this._camera;
-    const reach = stepThisFrame + WASD_CAMERA_RADIUS_METRES;
-
-    // Floor under the camera now (collision floor).
-    const floorNow = this._probe.collisionFloorAt(
-      camera.position.x,
-      camera.position.z
-    );
-    // Destination column floor.
-    const destX = camera.position.x + dirX * reach;
-    const destZ = camera.position.z + dirZ * reach;
-    const floorDest = this._probe.collisionFloorAt(destX, destZ);
-
-    // Forward ray: from the camera along the horizontal travel direction,
-    // length `reach`, first accepted solid-floor hit (the wall/façade/cliff
-    // ahead).
-    const forwardHit = this._forwardRayHit(dirX, dirZ, reach);
-
-    const outcome = classifyWasdStep({
-      floorNow: { y: floorNow.y, normal: floorNow.normal },
-      floorDest: { y: floorDest.y, normal: floorDest.normal },
-      forwardHit,
-      reach,
-      targetDir: { x: dirX, z: dirZ },
-      lastBlocked: !!this._lastWasdBlocked
-    });
-
-    return {
-      kind: outcome,
-      floorDestY: floorDest.y,
-      floorNowY: floorNow.y,
-      destFloorHit: floorDest.source !== 'cache'
-    };
-  }
-
-  // TASK-024 (2b): cast a forward (horizontal) ray of length `reach` along
-  // (dirX, 0, dirZ) and return the first accepted solid-floor hit as
-  // { hit, dist, normalY, normalH } — or { hit: false } when clear.
-  _forwardRayHit(dirX, dirZ, reach) {
-    const sceneEl = this._sceneEl;
-    if (!sceneEl || !sceneEl.object3D) return { hit: false };
-    this._tmpV3a.set(dirX, 0, dirZ);
-    if (this._tmpV3a.lengthSq() < 1e-9) return { hit: false };
-    this._tmpV3a.normalize();
-    this._raycaster.set(this._camera.position, this._tmpV3a);
-    this._raycaster.near = 0;
-    this._raycaster.far = reach;
-    const hits = this._raycaster.intersectObject(sceneEl.object3D, true);
-    for (const hit of hits) {
-      if (!isSolidFloorHit(hit)) continue;
-      const n = worldHitNormal(hit);
-      return {
-        hit: true,
-        dist: hit.distance,
-        normalY: n.y,
-        normalH: { x: n.x, z: n.z }
-      };
-    }
-    return { hit: false };
   }
 
   // --- LB screen-space pan (Stage 1 parity-plus) ---
